# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafePlayer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  admin: 'admin',
  viewer: 'viewer',
} as const;

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  role: RegisterInputRole;
  displayName?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface PinInput {
  /**
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
}

export interface SetPinInput {
  /**
     * @minLength 4
     * @maxLength 8
     */
  pin: string;
}

export interface PinVerifyResponse {
  valid: boolean;
  /** @nullable */
  unfilteredToken: string | null;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  viewer: 'viewer',
} as const;

export type UserTier = typeof UserTier[keyof typeof UserTier];


export const UserTier = {
  free: 'free',
  premium: 'premium',
} as const;

export interface User {
  id: number;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: UserRole;
  tier: UserTier;
  hasPin?: boolean;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserRoleUpdateRole = typeof UserRoleUpdateRole[keyof typeof UserRoleUpdateRole];


export const UserRoleUpdateRole = {
  admin: 'admin',
  viewer: 'viewer',
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export type MediaInputType = typeof MediaInputType[keyof typeof MediaInputType];


export const MediaInputType = {
  file: 'file',
  url: 'url',
} as const;

export interface MediaInput {
  type: MediaInputType;
  title?: string;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  fileHash?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  mimeType?: string | null;
}

export type MediaEntryType = typeof MediaEntryType[keyof typeof MediaEntryType];


export const MediaEntryType = {
  file: 'file',
  url: 'url',
} as const;

export type MediaEntrySafetyStatus = typeof MediaEntrySafetyStatus[keyof typeof MediaEntrySafetyStatus];


export const MediaEntrySafetyStatus = {
  safe: 'safe',
  unpreviewed: 'unpreviewed',
  flagged: 'flagged',
} as const;

export interface MediaEntry {
  id: number;
  type: MediaEntryType;
  title: string;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  fileHash?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  mimeType?: string | null;
  safetyStatus: MediaEntrySafetyStatus;
  jumpFrameCount?: number;
  /** @nullable */
  lastWatched?: string | null;
  createdAt: string;
}

export type MediaDetailSafetyStatus = typeof MediaDetailSafetyStatus[keyof typeof MediaDetailSafetyStatus];


export const MediaDetailSafetyStatus = {
  safe: 'safe',
  unpreviewed: 'unpreviewed',
  flagged: 'flagged',
} as const;

export type JumpFrameCategory = typeof JumpFrameCategory[keyof typeof JumpFrameCategory];


export const JumpFrameCategory = {
  violence: 'violence',
  sexual: 'sexual',
  language: 'language',
  other: 'other',
} as const;

export type JumpFrameSource = typeof JumpFrameSource[keyof typeof JumpFrameSource];


export const JumpFrameSource = {
  personal: 'personal',
  global: 'global',
} as const;

export interface JumpFrame {
  id: number;
  mediaId: number;
  /** seconds */
  startTime: number;
  /** seconds */
  endTime: number;
  category: JumpFrameCategory;
  source: JumpFrameSource;
  validated: boolean;
  /** @nullable */
  submittedBy?: number | null;
  createdAt: string;
}

/**
 * @nullable
 */
export type AiFlag = {
  flagged?: boolean;
  /** @nullable */
  reason?: string | null;
  severity?: 'low' | 'medium' | 'high' | 'extreme';
} | null;

export interface MediaDetail {
  media: MediaEntry;
  jumpFrames: JumpFrame[];
  safetyStatus: MediaDetailSafetyStatus;
  aiFlag?: AiFlag | null;
}

export interface MediaLookupInput {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  fileHash?: string | null;
}

export type MediaLookupResultSafetyStatus = typeof MediaLookupResultSafetyStatus[keyof typeof MediaLookupResultSafetyStatus];


export const MediaLookupResultSafetyStatus = {
  safe: 'safe',
  unpreviewed: 'unpreviewed',
  flagged: 'flagged',
} as const;

/**
 * @nullable
 */
export type MediaLookupResultSource = typeof MediaLookupResultSource[keyof typeof MediaLookupResultSource] | null;


export const MediaLookupResultSource = {
  personal: 'personal',
  global: 'global',
} as const;

export interface MediaLookupResult {
  found: boolean;
  /** @nullable */
  mediaId?: number | null;
  safetyStatus: MediaLookupResultSafetyStatus;
  jumpFrameCount?: number;
  /** @nullable */
  source?: MediaLookupResultSource;
}

export type JumpFrameInputCategory = typeof JumpFrameInputCategory[keyof typeof JumpFrameInputCategory];


export const JumpFrameInputCategory = {
  violence: 'violence',
  sexual: 'sexual',
  language: 'language',
  other: 'other',
} as const;

export interface JumpFrameInput {
  mediaId: number;
  startTime: number;
  endTime: number;
  category: JumpFrameInputCategory;
  submitToGlobal?: boolean;
}

export type SubmissionCategory = typeof SubmissionCategory[keyof typeof SubmissionCategory];


export const SubmissionCategory = {
  violence: 'violence',
  sexual: 'sexual',
  language: 'language',
  other: 'other',
} as const;

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  flagged: 'flagged',
} as const;

export interface Submission {
  id: number;
  mediaId: number;
  /** @nullable */
  mediaTitle?: string | null;
  startTime: number;
  endTime: number;
  category: SubmissionCategory;
  status: SubmissionStatus;
  submittedBy: number;
  /** @nullable */
  submitterEmail?: string | null;
  aiFlag?: AiFlag | null;
  createdAt: string;
}

export type FlaggedContentType = typeof FlaggedContentType[keyof typeof FlaggedContentType];


export const FlaggedContentType = {
  submission: 'submission',
  media: 'media',
} as const;

export type FlaggedContentSeverity = typeof FlaggedContentSeverity[keyof typeof FlaggedContentSeverity];


export const FlaggedContentSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  extreme: 'extreme',
} as const;

export interface FlaggedContent {
  id: number;
  type: FlaggedContentType;
  referenceId?: number;
  severity: FlaggedContentSeverity;
  reason: string;
  resolved?: boolean;
  createdAt: string;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  submission: 'submission',
  validation: 'validation',
  flag: 'flag',
  user_join: 'user_join',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  createdAt: string;
}

export interface AdminDashboard {
  totalUsers: number;
  totalMedia: number;
  pendingSubmissions: number;
  flaggedCount: number;
  globalFrameCount?: number;
  premiumUsers?: number;
  recentActivity: ActivityItem[];
}

export type ListJumpFramesParams = {
mediaId: number;
source?: ListJumpFramesSource;
};

export type ListJumpFramesSource = typeof ListJumpFramesSource[keyof typeof ListJumpFramesSource];


export const ListJumpFramesSource = {
  personal: 'personal',
  global: 'global',
  all: 'all',
} as const;

export type ListSubmissionsParams = {
status?: ListSubmissionsStatus;
};

export type ListSubmissionsStatus = typeof ListSubmissionsStatus[keyof typeof ListSubmissionsStatus];


export const ListSubmissionsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  flagged: 'flagged',
} as const;