# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SafePlayer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/verify-pin:
    post:
      operationId: verifyPin
      tags: [auth]
      summary: Verify admin PIN to unlock unfiltered mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinVerifyResponse"
        "401":
          description: Wrong PIN
  /auth/set-pin:
    post:
      operationId: setPin
      tags: [auth]
      summary: Set or update admin PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPinInput"
      responses:
        "200":
          description: OK

  # MEDIA
...
  /jumpframes:
    get:
      operationId: listJumpFrames
      tags: [jumpframes]
      summary: List jump frames for a given mediaId
      parameters:
        - name: mediaId
          in: query
          required: true
          schema:
            type: integer
        - name: source
          in: query
          required: false
          schema:
            type: string
            enum: [personal, global, all]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JumpFrame"

    post:
      operationId: createJumpFrame
      tags: [jumpframes]
      summary: Submit a new jump frame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JumpFrameInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JumpFrame"
  /jumpframes/{id}:
    delete:
      operationId: deleteJumpFrame
      tags: [jumpframes]
      summary: Delete a personal jump frame
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jumpframes/{id}/validate:
    post:
      operationId: validateJumpFrame
      tags: [jumpframes]
      summary: Admin validates a submitted jump frame into global database
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JumpFrame"

  # SUBMISSIONS
...
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, viewer]
        displayName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    PinInput:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          minLength: 4
          maxLength: 8
    SetPinInput:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          minLength: 4
          maxLength: 8
    PinVerifyResponse:
      type: object
      required: [valid, unfilteredToken]
      properties:
        valid:
          type: boolean
        unfilteredToken:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, role, tier]
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, viewer]
        tier:
          type: string
          enum: [free, premium]
        hasPin:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, viewer]

    # MEDIA SCHEMAS
...
    JumpFrame:
      type: object
      required: [id, mediaId, startTime, endTime, category, source, validated, createdAt]
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        startTime:
          type: number
          description: seconds
        endTime:
          type: number
          description: seconds
        category:
          type: string
          enum: [violence, sexual, language, other]
        source:
          type: string
          enum: [personal, global]
        validated:
          type: boolean
        submittedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    JumpFrameInput:
      type: object
      required: [mediaId, startTime, endTime, category]
      properties:
        mediaId:
          type: integer
        startTime:
          type: number
        endTime:
          type: number
        category:
          type: string
          enum: [violence, sexual, language, other]
        submitToGlobal:
          type: boolean

    # SUBMISSION SCHEMAS
...