# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SafePlayer API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserTier } from './userTier';

export interface User {
  id: number;
  email: string;
  /** @nullable */
  displayName?: string | null;
  role: UserRole;
  tier: UserTier;
  hasPin?: boolean;
  createdAt?: Date;
}